refactor: move plugin logic into lua/frenchcards/init.lua module
Extracts all logic from plugin/frenchcards.lua into a proper Lua module.
plugin/frenchcards.lua is reduced to a single require() call.
No behaviour change.
